+++ xen/include/asm-x86/domain.h
@@ -93,9 +93,9 @@ struct shadow_domain {
 
     /* Memory allocation */
     struct page_list_head freelist;
-    unsigned int      total_pages;  /* number of pages allocated */
-    unsigned int      free_pages;   /* number of pages on freelists */
-    unsigned int      p2m_pages;    /* number of pages allocates to p2m */
+    unsigned long      total_pages;  /* number of pages allocated */
+    unsigned long      free_pages;   /* number of pages on freelists */
+    unsigned long      p2m_pages;    /* number of pages allocates to p2m */
 
     /* 1-to-1 map for use when HVM vcpus have paging disabled */
     pagetable_t unpaged_pagetable;
@@ -155,9 +155,9 @@ struct shadow_vcpu {
 /************************************************/
 struct hap_domain {
     struct page_list_head freelist;
-    unsigned int      total_pages;  /* number of pages allocated */
-    unsigned int      free_pages;   /* number of pages on freelists */
-    unsigned int      p2m_pages;    /* number of pages allocates to p2m */
+    unsigned long      total_pages;  /* number of pages allocated */
+    unsigned long      free_pages;   /* number of pages on freelists */
+    unsigned long      p2m_pages;    /* number of pages allocates to p2m */
 };
 
 /************************************************/
